[refactorToRemoveAssetsFrontendComponents] Update the ErrorTemplate to GocukLibrary Twirl
