--- app/views/components/repeater_answer_section.scala.html
@@ -1,45 +1,0 @@
-@*
- * Copyright 2019 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import viewmodels.RepeaterAnswerSection
-@import views.html._
-
-@(answerSection: RepeaterAnswerSection)(implicit messages: Messages)
-
-<h2>@messages(answerSection.headingKey)</h2>
-
-<ul role="list" class="govuk-check-your-answers cya-questions-long section">
-    @components.answer_row(answerSection.relevanceRow)
-@for(row <- answerSection.rows){
-    <li>
-        <div class="cya-question">@row.answer</div>
-        <div class="cya-answer">
-            <a href="@row.deleteUrl">
-                <span aria-hidden="true">@messages("site.delete")</span>
-                <span class="visually-hidden">@messages("site.hidden-delete", row.answer)</span>
-            </a>
-        </div>
-        <div class="cya-change">
-            <a href="@row.changeUrl">
-                <span aria-hidden="true">@messages("site.edit")</span>
-                <span class="visually-hidden">@messages("site.hidden-edit", row.answer)</span>
-            </a>
-        </div>
-    </li>
-}
-</ul>
-
-<a href="@answerSection.addLinkUrl">@messages(answerSection.addLinkKey)</a>
